dev-requirements.txt
pytest-ckan
ckanext-scheming~=3.0.0
